Qfs C bindings: fix printf format specification.

+++ src/cc/qfsc/test-qfsc.c
@@ -258,7 +258,8 @@ static char* test_large_write() {
   check(strcmp(attr.filename, "file") == 0,
     "filename should be correct: %s != %s", attr.filename, "file");
   check(attr.size == len*2,
-    "file size should be correct: %li != %li", attr.size, len);
+    "file size should be correct: %li != %li",
+    (long long)attr.size, (long long)len);
 
   return 0;
 }
